requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart>=0.0.7
python-dotenv>=1.0.0
itsdangerous>=2.2.2
slowapi>=0.1.5